Address reviewer comments
