+++ components/support/nimbus-fml/src/cli.yaml
@@ -36,6 +36,16 @@ args:
 subcommands:
     - intermediate-repr:
         about: Generate the intermediate representation json, useful for debugging
+        # Uncomment this once we remove the global input and outputs above.
+        # args:
+        #     - INPUT:
+        #         help: Sets the input file to use
+        #         required: true
+        #         index: 1
+        #     - OUTPUT:
+        #         help: The file or directory where generated code is created
+        #         required: true
+        #         index: 2
     - generate:
         about: Generate feature structs against the Feature Variables API.
         args:
@@ -64,11 +74,11 @@ subcommands:
             # These next two flags aren't wired up yet.
             # This will be done in EXP-2573.
             - cache-dir:
-                help: The directory where downloaded files are cached
+                help: The directory where downloaded files are cached (inactive until EXP-2573)
                 long: cache-dir
                 takes_value: true
             - repo-file:
-                help: The file containing the version/refs/locations for other repos
+                help: The file containing the version/refs/locations for other repos (inactive until EXP-2573)
                 long: repo-file
                 takes_value: true
                 multiple: true
